// src/types/neutralino.d.ts
declare namespace Neutralino {

    namespace filesystem {
// ...
        interface Stats {
            size: number;
            isFile: boolean;
            isDirectory: boolean;
            createdAt: number;
            modifiedAt: number;
        }
// ...
    }
// ...
}